converges to infeasible point
Some changes to instant cost to check pressure/phi relationship.

Added new parameter: params.x_rest to distinguish resting parameters
from the initial condition of the optimization, params.x0

This converges to an infeasible point because x0 may not be compatible
with constraints. Going to see if eliminating phi as state fixes
anything